App works on emulator until this point

+++ MySecondApplication/app/src/main/java/com/example/mysecondapplication/MainActivity.kt
@@ -1,7 +1,10 @@
 package com.example.mysecondapplication
 
+import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.AttributeSet
+import android.view.View
 import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
@@ -11,12 +14,13 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val userInputValue = findViewById<EditText>(R.id.editTextNumber).text.toString().toInt()
-
+        val userInputValue = findViewById<EditText>(R.id.editTextNumber).text.toString()
+        //val userIpVal = findViewById<EditText>(R.id.editTextNumber).tex
+/*
         findViewById<Button>(R.id.button_check).setOnClickListener {
             Toast.makeText(this, "The entered number is $userInputValue !", Toast.LENGTH_LONG).show()
         }
-
+*/
     }
